backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
scipy==1.13.1
jax[cuda12]
opencv-python
opt-einsum>=2.3.2
safetensors>=0.4.3

